# packages/contracts/openapi/plan-ledger.yaml
openapi: 3.0.3
info:
  title: Full Shelf Plan Ledger API
  version: 1.0.0
  description: Deterministic mutation service, policy engine, and Spanner authority for Full Shelf.
security:
  - googleOidc: []
paths:
  /api/v1/approvals/approve-and-activate:
    post:
      summary: Independently verify the human operator, KMS-sign, persist approval, and activate
      parameters:
        - name: X-Full-Shelf-Operator-Authorization
          in: header
          required: true
          schema: { type: string }
      responses:
        '200': { description: Committed approval and activation receipt }
        '401': { description: Human Google ID token missing or invalid }
        '403': { description: Human operator not allowlisted }
        '409': { description: Revision or authoritative-state precondition failed }
        '503': { description: Identity or managed KMS failed closed }
  /api/v1/commands/execute:
    post:
      summary: Execute an authenticated, validated, idempotent ledger command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/ledger_command.json
      responses:
        '200':
          description: Stable transactional command receipt and replay metadata
        '409':
          description: Deterministic policy or state precondition rejected the command
  /api/v1/plans/preview:
    get:
      summary: Get morning distribution plan preview
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
            default: east-bay-food-bank
      responses:
        '200':
          description: Plan preview response
  /api/v1/actions/execute:
    post:
      deprecated: true
      summary: Retired legacy action route; use authenticated human approval route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action_id: { type: string }
                tenant_id: { type: string }
                agent_role: { type: string }
                action_type: { type: string }
                plan_id: { type: string }
                expected_revision: { type: string }
                parameters: { type: object }
                approval_envelope: { type: object }
                idempotency_key: { type: string }
      responses:
        '410':
          description: Authenticated human approval route required
  /api/v1/incidents/recall:
    post:
      deprecated: true
      summary: Legacy recall route; use the command endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lot_id: { type: string }
                hazard: { type: string }
                substitute_lot_id: { type: string }
      responses:
        '410':
          description: Complete deterministic command envelope required
  /api/v1/evidence/system:
    get:
      summary: Get System Evidence report for auditability
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: System Evidence drawer report
components:
  securitySchemes:
    googleOidc:
      type: http
      scheme: bearer
      bearerFormat: Google-signed OIDC ID token